Reorgainzed View
